Update Notes- EVF Received
I changed the dialog to remove the 1st one with the case number and footer month/year.
Added a dropdown menu to choose the EVF status for use of the case note header
I added the are_we_passworded out into the do loop. 
I removed the logic for the Complete or Incomplete header. 

Still needs to have the case note part updated with the new dialog information. The EVF status and date should be part of the header.
The tikl checkbox needs to be worked on in the code. Also need to remove the old Tikl stuff 
Needs to be tested.