Resolved #2715
Entered line to divide Verifs Requested from the rest of the CAF case note for counties that list the verifications needed at the top of the case note. 
Notes on income and budgeting error message changed to pop up at the end of the script. This eliminates the error message from interrupting when switching in between pages of the dialog.